changes to better describe functions.
I changed the text in the .R files a bit to make more clear the values used in function plastic_prev_plots  are obtained with the function plastic.ci. I don't know how to really update this in the manual, e.g. ?plastic.ci. Look like a roxy something is required.